Added New IK Prediction Method, and added IK foot adjustment to normal
Added a second function for predicting the IK target position which uses a default position. The target is predicted by using a vector starting from the endeffector, and going through the default position using a coefficient, the velocityPrediction.

Foot support added, that is the CCDIK class now has a toggle for using the last joint of a chain as a foot.  A foot will always adjust its orientation such that it is orthogonal to the normal of the surface it is standing on. The current implementation is still buggy though.