Added more Raycasts for IKStepper, added vertical cam rotation clamping to spider controller
IKStepper now uses a struct called Line for representation of Rays, since it is easier to  store origin and end point instead of origin and direction. Since the Ray class provided by unity does not store the distance and therefore not all the information necessary. Using my own struct it is easier to debug draw the rays using only the struct.

Added vertical clamping of the camera to the spider. Not perfect but it works. Made it such that the camera does not completely follow the spider when climbing walls but rather only rotates half the distance. Still need to make it so that the cam smoothly rotates to that desired position over a certain timeframe. Either by a coroutine or by lerping in the update.