Bug fixing and Refactoring
Added more efficient DrawSphere method to DebugShapes and DrawSphereRay.

Added foot angle option to IKSolver.

Added Sphereray option to IKStepper.

Added different camera clipping to SpiderController. Makes objects in vision sight invisible.

Fixed gravity of the spider, gravity now uses the ground normal and not the slowly lerping one the spider uses for orientation.

Refactored all the debug drawing of the spider using the new function in DebugShapes.

Added correctRotation to the IKSolver which should in theory not allow joints to go below the ground?