Refactored Raycasting to its own class Cast
Added the abstract class Cast which SphereCast and RayCast inherit from.
They represent rays with methods of casting implemented. This refactores all the  raycasting of the project to this class.

Added comfortable flag to TargetInfo instead of as out parameter of checkValidTarget.

Changed tolerance of IKSolver to 0.1f.

In the CCD solver the footangle is now only forced in the first half of iterations. After that the joint can be rotated normally.

Cleaned up Inspector of SpiderController and supplied it with titles.