Refactored all movements of the spider to FixedUpdate
Moved all movements and groundchecks of spider to fixed update since i want all movements of colliders to be done in fixed update.

Changed getInput() to project onto the current ground normal. This allows the spider to walk around the terrain insanely fast since i no longer move the spider to its current transform. (Since the spider might rotate to the normal slowly i was able to move in directions i shouldt be able to). Thix fixes the "running over cliffs" bug.

There is still a small bug where when the spider climbs walls it tilts slighly to the left or right?