Added LegCentroid and LegNormal adjustment. Added Breathing.
Spider now has references to its eight IK legs and the body of the rig.
The body admits the centroid position of the legs each frame and adjusts its rotation to the average normal. For now i deactivate these features since they dont work the way i want them to.