Refactored all Casts into Dictionary
Casts are now contained within a dictionary with a corresponding key.
This makes it easier to add rays and iterate through them.

The keys allow me to identify the casts and update their values if needed.

Moreover the ray that hit the target is now saved for debug purposes and then drawn in a different color.