Changes to Scaling
Set all scales (except for the main parent) to 1.
So, for scaling i simply use the main parents local y scale. So one has to make sure that the xyz scale coordinates are uniform.

Scaling still has a problem with the calculation of the default position, specifically the stride. This needs to be fixed. Moreover, scaling does not affect the IKSolvers tolerance or minimumChangePerIteration and this will affect the behavior significantly if the scaling is of high measure.

Changed names of the Scenes.

Added DrawLine and DrawRay to DebugShapes.