Fixed scaling of Spider
Default position of each IKChain now gets calculated correctly.
Refactored the tolerance and minimumChangePerIteration of the IKSolver to the IKChain which pass their values over when calling the CCD algorithm. This allowed me to multiply the values by the lossy scale such that both the tolerance and minimum change now scale with the spider.