Fixed Breathing and changes to tolerance
Added tolerance and minChange to the public variables of IKChain.
Display them as spheres at the endeffector for debug purposes.

Fixed breathing. Now uses the inititial body position parented to the spiders transform as pivot.
With breathing we definitely need a low solving tolerance for non jiggly movement. However a too low tolerance can lead to premature stepping.

Fixed Shader not being found in build by adding an object using a material using the transparent shader to the scene.