Singularity fix and Refactorization
Refactored the self initializations from Start to Awake.

Added singularity radius to the IKSolver, that is i skips a joint if the target is closer to it than the set radius.

SpiderNPCController now more prone to walk full speed and still standing since i applied a threshold to the speed.