Leg normal adjustment fixed and added drag to spider walk function
Renamed from TargetInfo the variable "comfortable" to "grounded".

Fixed leg normal adjustment. Now works fine.

Leg centroid adjustment sort of works but is not pretty.

Changed walk function of spider to have a drag. This was implemented in the SpiderController before but is now in the walk function itself.

Input function from SpiderController now supports analog input, that is the input isnt simply normalized but supports inputs with magnitude less than 1.