Breathing offset and leg normal changes
Breathing now offset such that the lowest breathing point is the start position.

Leg normal adjustment now uses the difference of orientation of the vector to the end effector and the tangent to the spiders normal. Prior i used the targets normal. Now the orientations adjusts itself whenever the leg rises and its therefore more natural.