Experimentation of new Step System
Implemented a StepManager which features a queue based step system as well as a alternating tetrapod gait system.

So instead of each leg checking whether it can step on its own, the manager acts as the brain which takes this decision from all of them.

The queue based system stores the legs that want to step in a queue and performs the stepping in the order of the queue. The system will wait if the top element can not step and this can inhibit the other legs from stepping for too long.

The alternating tetrapod gait system is inspired from how real life spiders walk. The legs are assigned one of two groups. Then a timer switches between these groups on a certain timeinterval. Every group only has a specific frame at which stepping is allowed in each timeinterval. With this the legs in the same group will always step at the same time if they need to step, and will never step while the other group is. One problem is spiders dont necessarily follow this gait at slow speeds.