Added FABRIK as an IK solving method (Only the basic framework)
Does not work yet:

Does not unparent childed joints when moving a joint which results in collateral movement of childed joints.

Does not set orientation of joint after translation, resulting in falsely oriented joints, which in turn breaks the rotation axis of the joint, since the axis defined in local coordinate space.

Does not include rotational limits yet.