Removed SwingTwistJoint
Removed the swing twist joint script since it is no longer used. This script is left over from a very early implementation, however I decided to restrict myself to a hinge joint implementation only. The reason for this is that hinges only have one degree of freedom making them a lot easier to handle and more performant. And in the case of a spider leg these degrees of freedom are plenty for the desired movement.